added date to expenses and factures, added export option, added multiplier to schedules

+++ IsoPlan/Data/Entities/ScheduleTotalPerEmployee.cs
@@ -3,7 +3,7 @@
     public class ScheduleTotalPerEmployee
     {
         public Employee Employee { get; set; }
-        public int TotalDays { get; set; }
+        public double TotalDays { get; set; }
         public double Salary { get; set; }
     }
 }
